pyproject.toml
[project]
name = "strands-verifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arize-otel>=0.10.0",
    "boto3>=1.40.36",
    "click>=8.3.0",
    "colorama>=0.4.6",
    "opentelemetry-sdk>=1.37.0",
    "prettytable>=3.16.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "strands-agents>=1.9.1",
]
